refactored notifications to linting standards
